# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "CohortMatrix": {
        "properties": {
          "as_of": {
            "format": "date",
            "title": "As Of",
            "type": "string"
          },
          "cohorts": {
            "items": {
              "$ref": "#/components/schemas/CohortRow"
            },
            "title": "Cohorts",
            "type": "array"
          },
          "weeks": {
            "items": {
              "type": "integer"
            },
            "title": "Weeks",
            "type": "array"
          }
        },
        "required": [
          "as_of",
          "weeks",
          "cohorts"
        ],
        "title": "CohortMatrix",
        "type": "object"
      },
...
      "Envelope_ActivationData_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/ActivationData"
          },
          "meta": {
            "$ref": "#/components/schemas/ResponseMeta"
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "title": "Envelope[ActivationData]",
        "type": "object"
      },
...
      "Envelope_CohortMatrix_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/CohortMatrix"
          },
          "meta": {
            "$ref": "#/components/schemas/ResponseMeta"
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "title": "Envelope[CohortMatrix]",
        "type": "object"
      },
      "Envelope_EngagementData_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/EngagementData"
          },
          "meta": {
            "$ref": "#/components/schemas/ResponseMeta"
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "title": "Envelope[EngagementData]",
        "type": "object"
      },
...
      "Envelope_ExperimentList_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/ExperimentList"
          },
          "meta": {
            "$ref": "#/components/schemas/ResponseMeta"
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "title": "Envelope[ExperimentList]",
        "type": "object"
      },
...
      "FieldError": {
        "properties": {
          "loc": {
            "items": {
              "type": "string"
            },
            "title": "Loc",
            "type": "array"
          },
          "msg": {
            "title": "Msg",
            "type": "string"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "FieldError",
        "type": "object"
      },
...
      "Problem": {
        "description": "RFC 9457 problem document (media type application/problem+json).",
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "string"
          },
          "errors": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/FieldError"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Errors"
          },
          "instance": {
            "title": "Instance",
            "type": "string"
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Id"
          },
          "status": {
            "title": "Status",
            "type": "integer"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "type": {
            "examples": [
              "urn:connecthub:problem:data-not-ready"
            ],
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "detail",
          "instance",
          "request_id"
        ],
        "title": "Problem",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Read-only business metrics from the ConnectHub analytics warehouse.\n\nEvery number comes from dbt models and Python analytics that the pipeline builds\nand Great Expectations validates; definitions are in `docs/metric-definitions.md`.\nErrors are RFC 9457 problem documents (`application/problem+json`).\n\nSuccessful data responses carry a weak `ETag`; send it back in `If-None-Match` to get\n`304 Not Modified` while the data is unchanged. Responses are cached per data version\n(the last validated pipeline run, `meta.data_version`).",
    "title": "ConnectHub Analytics API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/activation": {
      "get": {
        "description": "Source gold.fct_activation_daily / fct_activation_milestone_days, by signup date.\nDefault range: the 90 signup days ending at the last date with a complete 14-day window.",
        "operationId": "get_activation_api_activation_get",
        "parameters": [
          {
            "description": "Inclusive, YYYY-MM-DD. Defaults are relative to the last loaded day, not today.",
            "in": "query",
            "name": "start",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Inclusive, YYYY-MM-DD. Defaults are relative to the last loaded day, not today.",
              "title": "Start"
            }
          },
          {
            "description": "Inclusive, YYYY-MM-DD. Defaults are relative to the last loaded day, not today.",
            "in": "query",
            "name": "end",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Inclusive, YYYY-MM-DD. Defaults are relative to the last loaded day, not today.",
              "title": "End"
            }
          },
          {
            "description": "Billed plan tier filter",
            "in": "query",
            "name": "plan_tier",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "Free",
                    "Essentials",
                    "Professional",
                    "Enterprise"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Billed plan tier filter",
              "title": "Plan Tier"
            }
          },
          {
            "in": "query",
            "name": "granularity",
            "required": false,
            "schema": {
              "default": "week",
              "enum": [
                "day",
                "week",
                "month"
              ],
              "title": "Granularity",
              "type": "string"
            }
          },
          {
            "description": "Include signups whose 14-day window is not complete",
            "in": "query",
            "name": "include_incomplete",
            "required": false,
            "schema": {
              "default": false,
              "description": "Include signups whose 14-day window is not complete",
              "title": "Include Incomplete",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_ActivationData_"
                }
              }
            },
            "description": "Successful Response",
            "headers": {
              "Cache-Control": {
                "description": "private, max-age=<API_CACHE_MAX_AGE_S>",
                "schema": {
                  "type": "string"
                }
              },
              "ETag": {
                "description": "Weak validator of the response content",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "304": {
            "description": "Not Modified: If-None-Match matched the current ETag (no body)",
            "headers": {
              "Cache-Control": {
                "description": "private, max-age=<API_CACHE_MAX_AGE_S>",
                "schema": {
                  "type": "string"
                }
              },
              "ETag": {
                "description": "Weak validator of the response content",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {}
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {}
            },
            "description": "Unauthorized"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {}
            },
            "description": "Unprocessable Entity"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {}
            },
            "description": "Service Unavailable"
          },
          "504": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {}
            },
            "description": "Gateway Timeout"
          }
        },
        "summary": "14-day activation funnel",
        "tags": [
          "activation"
        ]
      }
    },
    "/api/cohorts": {
      "get": {
        "description": "Complete cells only; an ended week with no activity is 0; an unfinished week is null.",
        "operationId": "get_cohorts_api_cohorts_get",
        "parameters": [
          {
            "description": "Default: last loaded day",
            "in": "query",
            "name": "as_of",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Default: last loaded day",
              "title": "As Of"
            }
          },
          {
            "in": "query",
            "name": "cohort_start",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cohort Start"
            }
          },
          {
            "in": "query",
            "name": "cohort_end",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cohort End"
            }
          },
          {
            "in": "query",
            "name": "weeks",
            "required": false,
            "schema": {
              "default": 12,
              "maximum": 12,
              "minimum": 1,
              "title": "Weeks",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_CohortMatrix_"
                }
              }
            },
            "description": "Successful Response",
            "headers": {
              "Cache-Control": {
                "description": "private, max-age=<API_CACHE_MAX_AGE_S>",
                "schema": {
                  "type": "string"
                }
              },
              "ETag": {
                "description": "Weak validator of the response content",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "304": {
            "description": "Not Modified: If-None-Match matched the current ETag (no body)",
            "headers": {
              "Cache-Control": {
                "description": "private, max-age=<API_CACHE_MAX_AGE_S>",
                "schema": {
                  "type": "string"
                }
              },
              "ETag": {
                "description": "Weak validator of the response content",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {}
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {}
            },
            "description": "Unauthorized"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {}
            },
            "description": "Unprocessable Entity"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {}
            },
            "description": "Service Unavailable"
          },
          "504": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {}
            },
            "description": "Gateway Timeout"
          }
        },
        "summary": "Retention cohort matrix",
        "tags": [
          "retention"
        ]
      }
    },
...
    "/api/engagement": {
      "get": {
        "description": "Source gold.fct_daily_active_users. Default range: the 365 days ending at the last\nloaded day.",
        "operationId": "get_engagement_api_engagement_get",
        "parameters": [
          {
            "description": "Inclusive, YYYY-MM-DD. Defaults are relative to the last loaded day, not today.",
            "in": "query",
            "name": "start",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Inclusive, YYYY-MM-DD. Defaults are relative to the last loaded day, not today.",
              "title": "Start"
            }
          },
          {
            "description": "Inclusive, YYYY-MM-DD. Defaults are relative to the last loaded day, not today.",
            "in": "query",
            "name": "end",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Inclusive, YYYY-MM-DD. Defaults are relative to the last loaded day, not today.",
              "title": "End"
            }
          },
          {
            "in": "query",
            "name": "granularity",
            "required": false,
            "schema": {
              "default": "week",
              "enum": [
                "day",
                "week",
                "month"
              ],
              "title": "Granularity",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_EngagementData_"
                }
              }
            },
            "description": "Successful Response",
            "headers": {
              "Cache-Control": {
                "description": "private, max-age=<API_CACHE_MAX_AGE_S>",
                "schema": {
                  "type": "string"
                }
              },
              "ETag": {
                "description": "Weak validator of the response content",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "304": {
            "description": "Not Modified: If-None-Match matched the current ETag (no body)",
            "headers": {
              "Cache-Control": {
                "description": "private, max-age=<API_CACHE_MAX_AGE_S>",
                "schema": {
                  "type": "string"
                }
              },
              "ETag": {
                "description": "Weak validator of the response content",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {}
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {}
            },
            "description": "Unauthorized"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {}
            },
            "description": "Unprocessable Entity"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {}
            },
            "description": "Service Unavailable"
          },
          "504": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {}
            },
            "description": "Gateway Timeout"
          }
        },
        "summary": "DAU / WAU / MAU trend",
        "tags": [
          "engagement"
        ]
      }
    },
    "/api/experiments": {
      "get": {
        "description": "Registered experiments with their persisted evaluation (analytics.experiment_results).",
        "operationId": "get_experiments_api_experiments_get",
        "parameters": [
          {
            "in": "query",
            "name": "decision",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "SHIP",
                    "CONTINUE",
                    "HOLD",
                    "REVERT"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Decision"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "running",
                    "completed",
                    "not_evaluated"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_ExperimentList_"
                }
              }
            },
            "description": "Successful Response",
            "headers": {
              "Cache-Control": {
                "description": "private, max-age=<API_CACHE_MAX_AGE_S>",
                "schema": {
                  "type": "string"
                }
              },
              "ETag": {
                "description": "Weak validator of the response content",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "304": {
            "description": "Not Modified: If-None-Match matched the current ETag (no body)",
            "headers": {
              "Cache-Control": {
                "description": "private, max-age=<API_CACHE_MAX_AGE_S>",
                "schema": {
                  "type": "string"
                }
              },
              "ETag": {
                "description": "Weak validator of the response content",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {}
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {}
            },
            "description": "Unauthorized"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {}
            },
            "description": "Unprocessable Entity"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {}
            },
            "description": "Service Unavailable"
          },
          "504": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {}
            },
            "description": "Gateway Timeout"
          }
        },
        "summary": "Experiment portfolio",
        "tags": [
          "experiments"
        ]
      }
    },
...
  },
...
}